pyproject.toml
[project]
name = "syn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.27",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
